+++ src/extension.ts
@@ -2,7 +2,9 @@
 
 import * as fs from "fs";
 import * as vscode from "vscode";
-import { ServerOptions, Executable, LanguageClient, LanguageClientOptions } from 'vscode-languageclient';
+import { ServerOptions, Executable, LanguageClient, LanguageClientOptions, RequestType, TextDocumentPositionParams, Hover } from 'vscode-languageclient';
+import * as mime from "mime";
+import * as open from "open";
 import { Logger } from "./logger";
 import * as util from './common';
 import { ExtensionDownloader } from "./ExtensionDownloader";
@@ -81,6 +83,69 @@ export async function activate(context: vscode.ExtensionContext) {
     // client can be deactivated on extension deactivation
     context.subscriptions.push(client.start());
     context.subscriptions.push(restartServer);
+
+    // Register custom command to allow includes, literalincludes, and figures to be clickable
+    let hoverFile: string;
+    let clickInclude = vscode.commands.registerCommand('snooty.clickInclude', async () => {
+        // Send request to server (snooty-parser)
+        let type = mime.getType(hoverFile);
+
+        if (type == null || !type.includes("image")) {
+            let textDoc = await vscode.workspace.openTextDocument(hoverFile);
+            vscode.window.showTextDocument(textDoc);
+        }
+        else {
+            open(hoverFile);
+        }
+    });
+
+    context.subscriptions.push(clickInclude);
+
+    // Shows clickable link to file after hovering over it
+    vscode.languages.registerHoverProvider(
+        clientOptions.documentSelector,
+        new class implements vscode.HoverProvider {
+          provideHover(
+            _document: vscode.TextDocument,
+            _position: vscode.Position,
+            _token: vscode.CancellationToken
+          ): vscode.ProviderResult<vscode.Hover> {
+            // Get range for a link
+            let wordRegex = /\/\S+/;
+            let wordRange = _document.getWordRangeAtPosition(_position, wordRegex);
+
+            if (wordRange != undefined) {
+                // Get text at that range
+                let word = _document.getText(wordRange);
+
+                // Request hover information using the snooty-parser server
+                let request = async () => {
+                    let contents: vscode.MarkdownString;
+
+                    await client.sendRequest("textDocument/resolve", {path: word}).then((file: string) => {
+                        hoverFile = file;
+                        const command = vscode.Uri.parse(`command:snooty.clickInclude`);
+
+                        // Clean up absolute path for better UX. str.match() was not working with regex but can look into later
+                        let workspaceFolder = vscode.workspace.name;
+                        let folderIndex = hoverFile.search(workspaceFolder);
+                        let hoverPathRelative = hoverFile.slice(folderIndex);
+
+                        contents = new vscode.MarkdownString(`[${hoverPathRelative}](${command})`);
+                        contents.isTrusted = true; // Enables commands to be used
+                    });
+
+                    return new vscode.Hover(contents, wordRange);
+                }
+
+                return request();
+            }
+            else {
+                return new vscode.Hover(null);
+            }
+          }
+        } ()
+      );
 }
 
 // this method is called when your extension is deactivated